% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.getdbfile}
\alias{.getdbfile}
\title{Hidden function that creates a local copy of the database.}
\usage{
.getdbfile()
}
\arguments{
\item{query}{}
}
\value{
path to the newly created database
}
\description{
Hidden function that creates a local copy of the database.
}
\keyword{internal}
