% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Hidden load function.
Check the presence of the database and load it with AnnotationHub
if required.
TODO: add timestamp validation.}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{the path where the package is installed}

\item{pkgname}{name of the package}
}
\value{
dbConnect
}
\description{
Hidden load function.
Check the presence of the database and load it with AnnotationHub
if required.
TODO: add timestamp validation.
}
\keyword{internal}
