% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{findGenesByEntrez}
\alias{findGenesByEntrez}
\title{Find GeneIDs for Entrez}
\usage{
findGenesByEntrez(entrez)
}
\arguments{
\item{entrez}{Entres IDs}
}
\value{
data.frame
}
\description{
Find GeneIDs for Entrez
}
