% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getAllGenes4BrainRegion}
\alias{getAllGenes4BrainRegion}
\title{Get list of all genes found in particular BrainRegion for particular specie}
\usage{
getAllGenes4BrainRegion(brainRegion, taxID)
}
\arguments{
\item{taxID}{}
}
\value{

}
\description{
Get list of all genes found in particular BrainRegion for particular specie
}
