% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getAllGenes4Compartment}
\alias{getAllGenes4Compartment}
\title{Get list of all genes found in compartment}
\usage{
getAllGenes4Compartment(compartmentID)
}
\arguments{
\item{compartmentID}{}
}
\value{

}
\description{
Get list of all genes found in compartment
}
