% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_regions.R
\name{getBrainRegions}
\alias{getBrainRegions}
\title{Get list of all Brain regions in the database}
\usage{
getBrainRegions()
}
\value{
data.frame
}
\description{
Get list of all Brain regions in the database
}
