% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{getCompartments}
\alias{getCompartments}
\title{Provide full list of compartments (pres, post, synaptosome)}
\usage{
getCompartments()
}
\value{
data.frame
}
\description{
Provide full list of compartments (pres, post, synaptosome)
}
