% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getfGfPfR}
\alias{getfGfPfR}
\title{Get data from FullGeneFullPaperFullRegion view.}
\usage{
getfGfPfR(ids)
}
\arguments{
\item{ids}{}
}
\value{

}
\description{
Get data from FullGeneFullPaperFullRegion view.
}
