% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{getGeneDiseaseByEntres}
\alias{getGeneDiseaseByEntres}
\title{Get data about Gene-Disease mapping by Entres IDs}
\usage{
getGeneDiseaseByEntres(entrez)
}
\arguments{
\item{entrez}{entrez IDs}
}
\value{
data.frame
}
\description{
Get data about Gene-Disease mapping by Entres IDs
}
