% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGeneDiseaseByIDs}
\alias{getGeneDiseaseByIDs}
\title{Get data about Gene-Disease mapping by GeneIDs}
\usage{
getGeneDiseaseByIDs(ids)
}
\arguments{
\item{ids}{}
}
\value{

}
\description{
Get data about Gene-Disease mapping by GeneIDs
}
