% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{getGeneDiseaseByName}
\alias{getGeneDiseaseByName}
\title{Disease information for Human Gene Names}
\usage{
getGeneDiseaseByName(names)
}
\arguments{
\item{names}{\code{vector} of Human gene names}
}
\value{
data.frame
}
\description{
Get Human disease information (HDO provided) for the set of
Human Gene Names. Function lookups for diseases associated with
internal Human gene names and returns list of available diseases.
}
\examples{
t <- getGeneDiseaseByName (c('CASK', 'DLG2', 'DLG1')) #(115 rows)
}
\seealso{
Other {disease_functions}: 
\code{\link{getGeneDiseaseByEntres}()},
\code{\link{getGeneDiseaseByIDs}()}
}
\concept{{disease_functions}}
