% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{getGeneIdByCompartmentPaperCnt}
\alias{getGeneIdByCompartmentPaperCnt}
\title{Get list of frequently found in `Compartment GeneIDs}
\usage{
getGeneIdByCompartmentPaperCnt(cnt = 1)
}
\arguments{
\item{cnt}{minimal number of papers that mentioned gene}
}
\value{
tibble wiht GeneID, LocalisationID, and Npmid
        columns for genes and paper count
        data respectively.
}
\description{
Get list of frequently found in `Compartment GeneIDs
}
\examples{
cntT<-synaptome.db:::getGeneIdByCompartmentPaperCnt(4)
}
\keyword{internal}
