% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGeneIdByName}
\alias{getGeneIdByName}
\title{Get list of GeneIDs corresponding to provided gene names.}
\usage{
getGeneIdByName(name)
}
\arguments{
\item{name}{\code{vector} of gene names.}
}
\value{
\code{vector} of GeneID values.
}
\description{
Get list of GeneIDs corresponding to provided gene names.
}
\examples{
t <- getGeneIdByName(c('Src', 'Srcin1', 'Fyn'))
}
