% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{getGeneIdByPaperCnt}
\alias{getGeneIdByPaperCnt}
\title{Get list of frequently found GeneIDs}
\usage{
getGeneIdByPaperCnt(cnt = 1)
}
\arguments{
\item{cnt}{minimal number of papers that mentioned gene}
}
\value{
tibble wiht GeneID and Npmid columns for genes and paper count
        data respectively.
}
\description{
Get list of frequently found GeneIDs
}
\examples{
cntT<-synaptome.db:::getGeneIdByPaperCnt(47)
}
\keyword{internal}
