% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{getGeneIdByPapers}
\alias{getGeneIdByPapers}
\title{Get list of GeneIDs for genes found in specified papers}
\usage{
getGeneIdByPapers(pmids, cnt = 1)
}
\arguments{
\item{pmids}{vector of PMIDs to search for genes}

\item{cnt}{minimal number of papers that mentioned gene}
}
\value{
tibble wiht GeneID and Npmid columns for genes and paper count
        data respectively.
}
\description{
Get list of GeneIDs for genes found in specified papers
}
\examples{
res<- getAllGenes4Compartment(compartmentID = 1)
gnt<-getGeneInfoByIDs(res$GeneID)
pmids<-names(sort(table(gnt$PaperPMID))[1:5])
cntT<-synaptome.db:::getGeneIdByPapers(pmids,3)
}
\keyword{internal}
