% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGeneInfoByName}
\alias{getGeneInfoByName}
\title{Get gene information for set of gene names.}
\usage{
getGeneInfoByName(name)
}
\arguments{
\item{name}{\code{vector} of gene names}
}
\value{
\code{data.frame} with fields specified above
}
\description{
Function lookup for name in Human Gene name and Mouse Gene name data
}
\details{
This function then returns
following features for all found genes:
\itemize{
 \item GeneID,
 \item Localisation,
 \item MGI,
 \item HumanEntrez,
 \item MouseEntrez,
 \item HumanName,
 \item MouseName,
 \item PaperPMID,
 \item Paper,
 \item Year,
 \item SpeciesTaxID,
 \item BrainRegion
}
}
