% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getInducedPPI4BrainRegion}
\alias{getInducedPPI4BrainRegion}
\title{Prepare induces network for compartment}
\usage{
getInducedPPI4BrainRegion(ids, brainRegion, taxID)
}
\arguments{
\item{ids}{}
}
\value{

}
\description{
Prepare induces network for compartment
}
