% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getLimitedPPI4BrainRegion}
\alias{getLimitedPPI4BrainRegion}
\title{Prepare limited network for compartment}
\usage{
getLimitedPPI4BrainRegion(ids, brainRegion, taxID)
}
\arguments{
\item{ids}{}
}
\value{

}
\description{
Prepare limited network for compartment
}
