% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getLimitedPPI4Compartment}
\alias{getLimitedPPI4Compartment}
\title{Prepare limited network for compartment}
\usage{
getLimitedPPI4Compartment(ids, compartmentID)
}
\arguments{
\item{compartmentID}{}
}
\value{

}
\description{
Prepare limited network for compartment
}
