% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getMutations4DiseaseByEntres}
\alias{getMutations4DiseaseByEntres}
\title{Show mutations for specific disease by Entrez IDs}
\usage{
getMutations4DiseaseByEntres(entrez, hdoid)
}
\arguments{
\item{hdoid}{}
}
\value{

}
\description{
Show mutations for specific disease by Entrez IDs
}
