% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getMutations4DiseaseByName}
\alias{getMutations4DiseaseByName}
\title{Show mutations for specific disease by gene name}
\usage{
getMutations4DiseaseByName(name, hdoid)
}
\arguments{
\item{hdoid}{}
}
\value{

}
\description{
Show mutations for specific disease by gene name
}
