% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{getPapers}
\alias{getPapers}
\title{Get synaptome papers overview}
\usage{
getPapers()
}
\value{
tibble with following columns:
\itemset{
\item PaperPMID
\item SpeciesTaxID
\item BrainRegionID
\item LocalisationID
\item MethodID
\item Ngenes
\item Year
\item Name
\item BrainRegion
\item Localisation
}
}
\description{
Get synaptome papers overview
}
\examples{
p <- getPapers()
head(p)
}
