% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synaptome_db.R
\docType{package}
\name{synaptome.db}
\alias{synaptome.db-package}
\alias{synaptome.db}
\title{synaptome.db: programmatic access to the Synaptic proteome database}
\description{
The package obtains a local copy of the Synaptic proteome database
from AnnotationHub package  \code{\link[synaptome.data]{synaptome.data}}
and provides a set of utility R functions to query and analyse its content.
}
\references{
\insertRef{Sorokina:2021hl}{synaptome.db}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/lptolik/synaptome.db/issues/}
}

}
\author{
\strong{Maintainer}: Oksana Sorokina \email{oksana.sorokina@ed.ac.uk}

Authors:
\itemize{
  \item Anatoly Sorokin \email{lptolik@gmail.com}
  \item J. Douglas Armstrong
}

}
