% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_regions.R
\name{getBrainRegions}
\alias{getBrainRegions}
\title{Get list of all Brain regions in the database}
\usage{
getBrainRegions()
}
\value{
data.frame with following columns:
\itemize{
\item ID: Brain region internal ID
\item Name: name of the region
\item Description: extended description of the region
\item InterlexID
\item ParentID: ID of the containing brain region
}
}
\description{
Get full list of brain regions considered in the database. Returns
table with regions and their respective IDs.
}
\examples{
t <- getBrainRegions()
}
\seealso{
Other {BrainRegion functions}: 
\code{\link{getAllGenes4BrainRegion}()},
\code{\link{getGenes4BrainRegion}()},
\code{\link{getPPIbyIDs4BrainRegion}()}
}
\concept{{BrainRegion functions}}
