% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{getCompartments}
\alias{getCompartments}
\title{Get full list of compartments}
\usage{
getCompartments()
}
\value{
data.frame described above
}
\description{
Get list of synaptic compartments considered in the database.
Returns table with compartments: “Presynaptic”, “Postsynaptic”,
“Synaptosome”, and their internal IDs.
}
\examples{
c <- getCompartments()
}
\seealso{
Other {Compartment functions}: 
\code{\link{getAllGenes4Compartment}()},
\code{\link{getGenes4Compartment}()},
\code{\link{getPPIbyIDs4Compartment}()}
}
\concept{{Compartment functions}}
