% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGeneDiseaseByEntres}
\alias{getGeneDiseaseByEntres}
\title{Get data about Gene-Disease mapping by Entres IDs}
\usage{
getGeneDiseaseByEntres(entrez)
}
\arguments{
\item{ids}{}
}
\value{

}
\description{
Get data about Gene-Disease mapping by Entres IDs
}
