% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{getGeneDiseaseByIDs}
\alias{getGeneDiseaseByIDs}
\title{Disease information for internal gene IDs.}
\usage{
getGeneDiseaseByIDs(ids)
}
\arguments{
\item{ids}{\code{vector} of gene IDs.}
}
\value{
data.frame with following columns:
\itemize{
\item HumanEntrez
\item HumanName
\item HDOID
\item Description
}
}
\description{
Get Human disease information (HDO provided) for the set of internal
gene IDs. Function lookups for diseases associated with internal
GeneIDs and returns list of available diseases.
}
\examples{
t <- getGeneDiseaseByIDs (c(48, 585, 710))
}
\seealso{
Other {Disease functions}: 
\code{\link{getGeneDiseaseByEntres}()},
\code{\link{getGeneDiseaseByName}()}
}
\concept{{Disease functions}}
