% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGeneDiseaseByName}
\alias{getGeneDiseaseByName}
\title{Get data about Gene-Disease mapping by  Name.}
\usage{
getGeneDiseaseByName(names)
}
\arguments{
\item{names}{}
}
\value{

}
\description{
Get data about Gene-Disease mapping by  Name.
}
