% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGeneInfoByEntrez}
\alias{getGeneInfoByEntrez}
\title{Get GeneInfo table for set of Entres IDs}
\usage{
getGeneInfoByEntrez(entrez)
}
\arguments{
\item{entrez}{}
}
\value{

}
\description{
Get GeneInfo table for set of Entres IDs
}
