% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGeneInfoByIDs}
\alias{getGeneInfoByIDs}
\title{Get GeneInfo table for set of GeneIDs}
\usage{
getGeneInfoByIDs(ids)
}
\arguments{
\item{ids}{}
}
\value{

}
\description{
Get GeneInfo table for set of GeneIDs
}
