% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGeneInfoByName}
\alias{getGeneInfoByName}
\title{Get gene information for set of gene names.
Function lookup for name in Human Gene name and Mouse Gene name data and return
following features for all found genes:
GeneID,
Localisation,
MGI,
HumanEntrez,
MouseEntrez,
HumanName,
MouseName,
PMID,
Paper,
Year,
SpeciesTaxID,
BrainRegion}
\usage{
getGeneInfoByName(name)
}
\arguments{
\item{name}{vector of gene names}
}
\value{
data.frame with fields specified above
}
\description{
Get gene information for set of gene names.
Function lookup for name in Human Gene name and Mouse Gene name data and return
following features for all found genes:
GeneID,
Localisation,
MGI,
HumanEntrez,
MouseEntrez,
HumanName,
MouseName,
PMID,
Paper,
Year,
SpeciesTaxID,
BrainRegion
}
