% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGenes4BrainRegion}
\alias{getGenes4BrainRegion}
\title{Select genes from list that found in particular BrainRegion for particular specie}
\usage{
getGenes4BrainRegion(ids, brainRegion, taxID)
}
\arguments{
\item{ids}{gene IDs}

\item{brainRegion}{region ID}

\item{taxID}{taxon ID}
}
\value{
data.frame
}
\description{
Select genes from list that found in particular BrainRegion for particular specie
}
