% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGenes4Compartment}
\alias{getGenes4Compartment}
\title{Select genes from list that found in compartment}
\usage{
getGenes4Compartment(ids, compartmentID)
}
\arguments{
\item{ids}{Gene IDs}

\item{compartmentID}{compartment ID}
}
\value{
data.frame
}
\description{
Select genes from list that found in compartment
}
