% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppiTables.R
\name{getIGraphFromPPI}
\alias{getIGraphFromPPI}
\title{Get Igraph representation of PPI}
\usage{
getIGraphFromPPI(ppi)
}
\arguments{
\item{ppi}{PPI \code{data.frame} with columns A and B, obtaioed
from functions like \code{\link{getPPIbyName}}}
}
\value{

}
\description{
Get Igraph representation of PPI
}
\examples{

}
