% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_regions.R
\name{getInducedPPI4BrainRegion}
\alias{getInducedPPI4BrainRegion}
\title{Prepare induces network for compartment}
\usage{
getInducedPPI4BrainRegion(ids, brainRegion, taxID)
}
\arguments{
\item{ids}{gene IDs}

\item{brainRegion}{region ID}

\item{taxID}{taxon ID}
}
\value{
tbl_lazy
}
\description{
Prepare induces network for compartment
}
