% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getInducedPPI4Compartment}
\alias{getInducedPPI4Compartment}
\title{Prepare induces network for compartment}
\usage{
getInducedPPI4Compartment(ids, compartmentID)
}
\arguments{
\item{compartmentID}{}
}
\value{

}
\description{
Prepare induces network for compartment
}
