% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_regions.R
\name{getLimitedPPI4BrainRegion}
\alias{getLimitedPPI4BrainRegion}
\title{Prepare limited network for compartment}
\usage{
getLimitedPPI4BrainRegion(ids, brainRegion, taxID)
}
\arguments{
\item{ids}{gene IDs}

\item{brainRegion}{region ID}

\item{taxID}{taxon ID}
}
\value{
tbl_lazy
}
\description{
Prepare limited network for compartment
}
