% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{getLimitedPPI4Compartment}
\alias{getLimitedPPI4Compartment}
\title{Prepare limited network for compartment}
\usage{
getLimitedPPI4Compartment(ids, compartmentID)
}
\arguments{
\item{ids}{Gene IDs}

\item{compartmentID}{compartment ID}
}
\value{
tbl_lazy
}
\description{
Prepare limited network for compartment
}
