% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{getMutations4DiseaseByEntres}
\alias{getMutations4DiseaseByEntres}
\title{Show mutations for specific disease by Entrez IDs}
\usage{
getMutations4DiseaseByEntres(entrez, hdoid)
}
\arguments{
\item{entrez}{Entrez IDs}

\item{hdoid}{HDO IDs}
}
\value{
data.frame
}
\description{
Show mutations for specific disease by Entrez IDs
}
