% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{getMutations4DiseaseByIDs}
\alias{getMutations4DiseaseByIDs}
\title{Show mutations for specific disease by GeneIDs}
\usage{
getMutations4DiseaseByIDs(ids, hdoid)
}
\arguments{
\item{ids}{Gene IDs}

\item{hdoid}{HDO IDs}
}
\value{
data.frame
}
\description{
Show mutations for specific disease by GeneIDs
}
