% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{getMutations4DiseaseByName}
\alias{getMutations4DiseaseByName}
\title{Show mutations for specific disease by gene name}
\usage{
getMutations4DiseaseByName(name, hdoid)
}
\arguments{
\item{name}{Gene name}

\item{hdoid}{HDO IDs}
}
\value{
data.frame
}
\description{
Show mutations for specific disease by gene name
}
