% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{getMutDiseaseQuery}
\alias{getMutDiseaseQuery}
\title{Prepare mutation access query}
\usage{
getMutDiseaseQuery()
}
\value{
\code{\link[dbplyr]{tbl_sql}} of table join
}
\description{
Prepare mutation access query
}
\keyword{internal}
