% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{getMutDiseaseQuery}
\alias{getMutDiseaseQuery}
\title{Prepare query for AllMutationAllPapers table}
\usage{
getMutDiseaseQuery()
}
\value{
tbl_lazy
}
\description{
Prepare query for AllMutationAllPapers table
}
