% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{getPapers}
\alias{getPapers}
\title{Get synaptome papers overview}
\usage{
getPapers()
}
\value{
tibble with following columns:
\itemize{
\item PaperPMID
\item SpeciesTaxID
\item Year
\item Name
\item Localisation
\item BrainRegion
\item Method
\item Ngenes
}
}
\description{
Get synaptome papers overview
}
\examples{
p <- getPapers()
head(p)
}
