% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppiTables.R
\name{getPPIbyEntrez}
\alias{getPPIbyEntrez}
\title{Extract the PPIs for my list of genes defined by Entrez IDs}
\usage{
getPPIbyEntrez(entrez, type = c("induced", "limited"))
}
\arguments{
\item{entrez}{Entrez IDs}

\item{type}{type of the PPI network}
}
\value{
data.frame
}
\description{
Extract the PPIs for my list of genes defined by Entrez IDs
}
