% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getPPIbyIDs}
\alias{getPPIbyIDs}
\title{Extract the PPIs for my list of genes defined by GeneID}
\usage{
getPPIbyIDs(ids, type = c("induced", "limited"))
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Extract the PPIs for my list of genes defined by GeneID
}
