% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppiTables.R
\name{getPPIbyIDs}
\alias{getPPIbyIDs}
\title{Extract the PPIs for my list of genes defined by GeneID}
\usage{
getPPIbyIDs(ids, type = c("induced", "limited"))
}
\arguments{
\item{ids}{Gene IDs}

\item{type}{type of the PPI network}
}
\value{
data.frame
}
\description{
Extract the PPIs for my list of genes defined by GeneID
}
\examples{
t <- getPPIbyIDs(c(48, 585, 710), type='limited') #(16 rows)
t <- getPPIbyIDs(c(48, 585, 710), type='induced') #306 rows
}
