% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_regions.R
\name{getPPIbyIDs4BrainRegion}
\alias{getPPIbyIDs4BrainRegion}
\title{Prepare induced or limited network for brain region}
\usage{
getPPIbyIDs4BrainRegion(
  ids,
  brainRegion,
  taxID,
  type = c("induced", "limited")
)
}
\arguments{
\item{ids}{gene IDs}

\item{brainRegion}{region ID}

\item{taxID}{taxon ID}

\item{type}{type of the PPI network}
}
\value{
data.frame
}
\description{
Prepare PPI network for genes from your list annotated
for specific brain region. Should be used with
[findGenesByEntrez()])
or [findGenesByName()]
functions to obtain list of internal IDs for your list of genes.
Could be used with getAllGenes4BrainRegion functions to obtain
all genes belonging to respective brain region. Function lookups
the PPI table for gene IDs from the list and returns “limited”
or “induced” interactors GeneIDs for the specified region.
}
