% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getPPIbyIDs4BrainRegion}
\alias{getPPIbyIDs4BrainRegion}
\title{Extract the PPIs for specific compartment}
\usage{
getPPIbyIDs4BrainRegion(
  ids,
  brainRegion,
  taxID,
  type = c("induced", "limited")
)
}
\arguments{
\item{taxID}{}
}
\value{

}
\description{
Extract the PPIs for specific compartment
}
