% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getPPIbyIDs4Compartment}
\alias{getPPIbyIDs4Compartment}
\title{Extract the PPIs for specific compartment}
\usage{
getPPIbyIDs4Compartment(ids, compartmentID, type = c("induced", "limited"))
}
\arguments{
\item{ids}{gene IDs}

\item{compartmentID}{compartment ID}

\item{type}{type of the PPI network}
}
\value{
data.frame
}
\description{
Extract the PPIs for specific compartment
}
