% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getPPIbyName}
\alias{getPPIbyName}
\title{Extract the PPIs for my list of genes defined by Gene name}
\usage{
getPPIbyName(name, type = c("induced", "limited"))
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Extract the PPIs for my list of genes defined by Gene name
}
