% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppiTables.R
\name{getTableFromPPI}
\alias{getTableFromPPI}
\title{Get table representation of the PPI.}
\usage{
getTableFromPPI(ppi)
}
\arguments{
\item{ppi}{PPI \code{data.frame} with columns A and B, obtaioed
from functions like \code{\link{getPPIbyName}}}
}
\value{
table with the following columns added for both interactors (A and B):
\itemize{
\item{GeneID internal database ID}
\item{MGI MGI ID}
\item{HumanEntrez Human Entrez ID}
\item{MouseEntrez Mouse Entrez ID}
\item{HumanName Human gene name}
\item{MouseName Mouse gene name}
\item{RatEntrez Rat Entrez ID}
\item{Rat Name Rat gene name}
}
}
\description{
Get table representation of the PPI.
}
\examples{
tbl<-getTableFromPPI(getPPIbyIDs(c(48, 585, 710), type='limited'))
}
