% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graphFromSynaptomeByEntrez}
\alias{graphFromSynaptomeByEntrez}
\title{Utility function to create network from
\code{\link[synaptome.db]{synaptome.db}} data}
\usage{
graphFromSynaptomeByEntrez(entrez, LCC = TRUE, simplify = TRUE)
}
\arguments{
\item{entrez}{vector of EntrezIDs for network vertices}

\item{LCC}{if TRUE only largest connected component is returned}

\item{simplify}{if TRUE loops and multiple edges will be removed}
}
\value{
network defined by the gene table
}
\description{
Utility function to create network from
\code{\link[synaptome.db]{synaptome.db}} data
}
\examples{
library(synaptome.db)
cid<-match('Presynaptic', getCompartments()$Name)
geneTable<-getAllGenes4Compartment(cid)
gg<-graphFromSynaptomeByEntrez(geneTable$HumanEntrez)
}
