% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graphFromSynaptomeGeneTable}
\alias{graphFromSynaptomeGeneTable}
\title{Utility function to create network from
\code{\link{synaptome.db}} data}
\usage{
graphFromSynaptomeGeneTable(geneTable, LCC = TRUE, simplify = TRUE)
}
\arguments{
\item{geneTable}{data.frame described in
\code{\link{getGenesByID}}}

\item{LCC}{if TRUE only largest connected component is returned}

\item{simplify}{if TRUE loops and multiple edges will be removed}
}
\value{
network defined by the gene table
#importFrom BioNAR buildNetwork
}
\description{
Utility function to create network from
\code{\link{synaptome.db}} data
}
\examples{
library(synaptome.db)
cid<-match('Presynaptic', getCompartments()$Name)
geneTable<-getAllGenes4Compartment(cid)
gg<-graphFromSynaptomeGeneTable(geneTable)
}
