% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synaptome_db.R
\docType{package}
\name{synaptome.db}
\alias{synaptome.db}
\alias{synaptome.db-package}
\title{synaptome.db: programmatic access to the Synaptic proteome database}
\description{
The package contains local copy of the Synaptic proteome database together
with a set of utility R functions to query and analyse its content.
}
\references{
\insertRef{Sorokina:2021hl}{synaptome.db}
}
\author{
\strong{Maintainer}: Oksana Sorokina \email{oksankasster@gmail.com}

Authors:
\itemize{
  \item Anatoly Sorokin \email{lptolik@gmail.com}
}

}
